backend/air_quality_service/requirements.txt
fastapi==0.115.11
slowapi==0.1.8
firebase_admin==6.7.0
protobuf==4.23.4
pydantic==2.10.6
uvicorn==0.34.0
google-cloud-firestore==2.20.0
google-auth==2.38.0
google-api-python-client==2.135.0
setuptools>=68.0.0
wheel>=0.37.0
google-api-core==2.24.2